requirements.txt
pyttsx3
SpeechRecognition
wikipedia
pywhatkit
requests
python-decouple
transformers
torch
